fifth

+++ src/main/java/com/pg/repository/DistrictRepository.java
@@ -6,5 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.pg.entity.District;
 
 @Repository
-public interface DistrictRepository extends JpaRepository<District, Long>{
+public interface DistrictRepository extends JpaRepository<District, Integer>{
 }
+
